Changed NameID and PropID from long to int in mAccount. Changed RenderedCode, PropertyType,Protest and SortCode from type char to type string in mMarionImport. Changed InterestType from type char to int in mMarionImport. Built mMarionImportRetrieve. Fixed bug in vmAccount line 185. Commented out owner.CadID = "MAR" in vmAgentAndOperator. Commented out UpdateRrcOperId in vmLease Added where RrcDist = '06' line 286 vmLease. Fixed bugs in vmMain that were preventing Dapper in Marion  Export. Added null check for agentCadOwnerID in vmMain. Commented out code to reverse out Account.SeqNmbr.

+++ Models/mAccount.cs
@@ -14,8 +14,8 @@ namespace MarionUpload.Models
         public float PctProp { get; set; }
         public bool Protest_YN { get; set; }
         public string PTDcode { get; set; }
-        public long NameID { get; internal set; }
-        public long PropID { get; internal set; }
+        public int NameID { get; internal set; }
+        public int PropID { get; internal set; }
         public string Cad { get; set; }
         public bool Stat_YN { get; set; }
         public string UpdateBy { get; set; }
